refactor(tooling): group repository checks by capability
