+++ package.json
@@ -18,11 +18,11 @@
     "bump-deps": "node scripts/dependencies/bump.ts && turbo run bump-deps --filter=cas",
     "bump-ui": "pnpm dlx shadcn@latest add --all --overwrite -c packages/design-system",
     "changeset": "changeset",
-    "check:patches": "node scripts/check-patches.ts",
+    "check:patches": "node scripts/check/patches.ts",
     "check:site": "pnpm --filter www exec vitest run --config vitest.afdocs.mts --coverage.enabled=false",
-    "check:tests": "node scripts/check-tests.ts",
+    "check:tests": "node scripts/check/tests.ts",
     "ci:gate": "node scripts/github/queue/main.ts gate",
-    "ci:production-acceptance": "node scripts/production-acceptance.ts",
+    "ci:production-acceptance": "node scripts/production/acceptance.ts",
     "ci:queue": "node scripts/github/queue/main.ts admit",
     "ci:review": "node scripts/github/queue/main.ts review",
     "clean": "rm -rf node_modules apps/*/node_modules packages/*/node_modules .cache apps/*/.cache packages/*/.cache .pnpm-store apps/*/.turbo packages/*/.turbo apps/*/.ruff_cache apps/*/.pytest_cache apps/*/.venv apps/*/.coverage .turbo .next apps/*/.next apps/*/.react-email packages/*/.react-email .vercel/output apps/*/.vercel/output coverage apps/*/coverage packages/*/coverage && find apps packages -name __pycache__ -type d -prune -exec rm -rf {} +",
@@ -34,8 +34,8 @@
     "dev:all": "turbo run dev",
     "dev:web": "turbo run dev --filter=www --filter=@repo/backend --filter=cas",
     "doctor": "pnpm --filter www run doctor",
-    "effect:source:check": "node scripts/effect-source.ts check",
-    "effect:source:update": "node scripts/effect-source.ts update",
+    "effect:source:check": "node scripts/effect/source.ts check",
+    "effect:source:update": "node scripts/effect/source.ts update",
     "format": "ultracite fix && turbo run format --filter=cas",
     "format:doctor": "ultracite doctor",
     "lint": "pnpm effect:source:check && pnpm check:patches && pnpm check:tests && ultracite check && turbo run lint --filter=cas",
@@ -44,7 +44,7 @@
     "start": "turbo run start",
     "test": "pnpm check:patches && pnpm check:tests && pnpm test:scripts && turbo run test",
     "test:coverage": "pnpm check:patches && pnpm check:tests && turbo run test:coverage",
-    "test:scripts": "pnpm typecheck:scripts && pnpm --filter @repo/testing exec vitest run --root ../.. scripts/dependencies/policy.test.ts scripts/github/cli.test.ts scripts/github/policy.test.ts scripts/github/provenance/bundle.test.ts scripts/github/provenance/verify.test.ts scripts/github/queue/admission.test.ts scripts/github/queue/release.test.ts scripts/github/queue/remote.test.ts scripts/github/queue/replay.test.ts scripts/github/queue/result.test.ts scripts/github/queue/tree.test.ts scripts/github/release.test.ts scripts/vercel/scope.test.ts scripts/effect-source.test.ts scripts/production-acceptance.test.ts",
+    "test:scripts": "pnpm typecheck:scripts && pnpm --filter @repo/testing exec vitest run --root ../.. scripts/dependencies/policy.test.ts scripts/github/cli.test.ts scripts/github/policy.test.ts scripts/github/provenance/bundle.test.ts scripts/github/provenance/verify.test.ts scripts/github/queue/admission.test.ts scripts/github/queue/release.test.ts scripts/github/queue/remote.test.ts scripts/github/queue/replay.test.ts scripts/github/queue/result.test.ts scripts/github/queue/tree.test.ts scripts/github/release.test.ts scripts/vercel/scope.test.ts scripts/effect/source.test.ts scripts/production/acceptance.test.ts",
     "test:ui": "turbo run test:ui",
     "test:watch": "turbo run test:watch",
     "translate": "turbo run translate",
