+++ scripts/check/patches.ts
